refactor(#209): move chat message and mention events to shared contracts
