+++ src/Services/Chat/ChatService.Api/Application/Messages/EditMessageService.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.Options;
+using Urfu.Link.BuildingBlocks.Contracts.Integration.Chat;
 using Urfu.Link.Services.Chat.Application.Contracts;
 using Urfu.Link.Services.Chat.Application.Mentions;
 using Urfu.Link.Services.Chat.Domain.Aggregates;
